move code around, add some more string functions
